Added boxy M4 model for Gun1, some issues persist with swapping from the red gun to the m4 (colliders)
i've unchecked isTrigger on the box collider for the gun. When isTrigger is selected, the gun will move on its transform axis out of the players hand and teleport somewhere random in the map. When isTrigger isn't selected, you will not be able to pick it up off the ground. Headed to bed now but I will look more into this tomorrow.