Added the Unity-chan character model
Initially there were a bunch of warnings generated (and 1 error). I corrected the error and fixed a few of the warnings. But after I ran the game to test, the remaining warnings disappeared and I cannot find a way to recreate them. Nothing depends on them so it should be ok for now but I would like to clear those warnings if possible.