Fixed bug
Before this commit, when user logged in and the camera moved the from one position to the other it would occasionally not stop moving. Fixedy-fixed that. I deserve an award. I'm basically a hero. No one is ever going to read this. I'm just talking to myself. Butter is better than batter but Betty bets batter beats butter. Say that 10 times fast.