Enabled scene cameras when leave room button is clicked
Before this fix, when a user tried to leave the room, there would be a second or two where there were no active cameras because their player's camera was destroyed and the new scene's cameras weren't loaded yet. Now, when a user clicks the leave room button, the scene cameras are turned back on so it doesn't look like the game froze/broke.