CameraAnimationHandler.MoveRight() --> now resets the stats panel text
I also added a namespace to PlayFabController, and added a null check for username and password (and status panel text to accompany it).

I tried to "break" this via monkey testing and was genuinely having trouble getting it to produce an unhanded error or undefined behavior.

I also started modifying the error messages to prevent sensitive data exposure. Not 100% sure why all of them are not being replaced atm but there are more pressing things to jump into now.