Wall health regen
I made the regen rate independent of the wall's max/original health. In other words, if regen rate is 5 that's like saying the wall will regain 5% of its original health every second. This way, when we want to figure out how strong we want the wall to be, we don't have to readjust the the regen rate to compensate for a stronger/weaker wall.