Wall vote implementation
Now every player gets to vote on the wall coming down by shooting it once (and only once).

There is a threshold where >50% of the players on one team need to vote for it to come down.

The wall is visually divided into two sides (Team A side and Team B side) and the wall's health is reflected in its color. The color represents the proportion of the number of players who still need to "vote" for the wall to come down.

For example, if there are 3 players on one team, 2 of them need to "vote" for the wall to come down. If 1 player "votes" the wall's health will be at 50%.

Each team can only see the "votes" their team cast. In other words, Team A shoots the wall and Team B will see no change in the color of their side of the wall. This allows one team to surprise another team by bringing the wall down unexpectedly.